requirements.txt
autopep8==1.6.0
azure-common==1.1.27 
azure-nspkg==3.0.2   
azure-storage==0.31.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
crypto==1.4.1
cryptography==35.0.0
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Naked==0.1.31
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0